% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_indices.R
\name{simpson}
\alias{simpson}
\title{Simpson Diversity Index}
\usage{
simpson(community, type = c("gini_simpson", "inverse", "dominance"))
}
\arguments{
\item{community}{A numeric vector of species abundances.}

\item{type}{One of \code{"inverse"} (1/D), \code{"gini_simpson"} (1 - D),
or \code{"dominance"} (D). Default is \code{"gini_simpson"}.}
}
\value{
A numeric value representing the Simpson index.
}
\description{
Calculates the Simpson diversity index (1 - D) for a community.
}
\details{
Simpson's dominance index D is calculated as:
\deqn{D = \sum_{i=1}^{S} p_i^2}
The Gini-Simpson index is \eqn{1 - D} and the inverse Simpson is
\eqn{1/D}.
}
\examples{
comm <- c(10, 5, 8, 3, 12)
simpson(comm)
simpson(comm, type = "inverse")

}
\seealso{
\code{\link[=shannon]{shannon()}} for Shannon diversity.
}
