% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_indices.R
\name{shannon}
\alias{shannon}
\title{Shannon Diversity Index}
\usage{
shannon(community, base = exp(1))
}
\arguments{
\item{community}{A numeric vector of species abundances.}

\item{base}{The logarithm base. Default is \code{exp(1)} (natural log).
Use \code{2} for bits.}
}
\value{
A numeric value representing the Shannon diversity index.
}
\description{
Calculates the Shannon-Wiener diversity index (H') for a community.
}
\details{
The Shannon index is calculated as:
\deqn{H' = -\sum_{i=1}^{S} p_i \ln(p_i)}
where \eqn{p_i} is the proportion of species \eqn{i} and \eqn{S}
is the total number of species.
}
\examples{
comm <- c(10, 5, 8, 3, 12)
shannon(comm)

}
