% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdiv-package.R
\docType{package}
\name{taxdiv-package}
\alias{taxdiv}
\alias{taxdiv-package}
\title{taxdiv: Taxonomic Diversity Indices Using Deng Entropy}
\description{
Calculates taxonomic diversity indices for ecological community data using Deng entropy framework and classical approaches (Shannon, Simpson, Clarke & Warwick). Provides functions for computing taxonomic distinctness, average taxonomic distinctness (AvTD/Delta+), variation in taxonomic distinctness (VarTD/Lambda+), and Deng entropy-based measures that incorporate taxonomic hierarchy information. Includes tools for constructing taxonomic trees, computing pairwise taxonomic distances, and visualizing diversity patterns across ecological gradients.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mgorgoz/taxonomic-diversity-r}
  \item Report bugs at \url{https://github.com/mgorgoz/taxonomic-diversity-r/issues}
}

}
\author{
\strong{Maintainer}: Murat Gorgoz \email{muratgorgoz@ogm.gov.tr}

}
\keyword{internal}
